fix(review): move prompt builder to franchise, detect_prefix to core, fix sys.modules bypass
Resolves cross-L3 server→cli coupling in _get_food_truck_prompt_builder:
- Move _build_food_truck_prompt/_build_l2_sous_chef_block to franchise._prompts
- Move detect_autoskillit_mcp_prefix/DIRECT_PREFIX/MARKETPLACE_PREFIX to core._plugin_ids
- server/helpers.py now imports from autoskillit.core and autoskillit.franchise (no cli)
- Replace sys.modules["autoskillit.execution"] lookup in _run_dispatch with direct import
- Add stable franchise_internal_error code to execute_dispatch catch-all handler
