+++ src/autoskillit/cli/_mcp_names.py
@@ -1,46 +1,23 @@
-"""MCP tool name prefix detection for CLI prompt builders.
+"""MCP tool name prefix detection and resolution for the CLI.
 
-Detects whether autoskillit is marketplace-installed or running under
-direct --plugin-dir only, and derives the correct fully-qualified MCP
-tool name prefix. Detection is pure Python I/O — no LLM, no subprocess,
-no network calls.
+Re-exports core detection primitives and provides resolve_tool_name for
+constructing fully-qualified tool names within CLI prompt builders.
 """
 
 from __future__ import annotations
 
-import json
-from pathlib import Path
+from autoskillit.core import (
+    DIRECT_PREFIX,
+    MARKETPLACE_PREFIX,
+    detect_autoskillit_mcp_prefix,
+)
 
-# The key written to installed_plugins.json by `autoskillit install`
-_AUTOSKILLIT_PLUGIN_KEY = "autoskillit@autoskillit-local"
-
-# Single source of truth for both known prefix forms
-DIRECT_PREFIX = "mcp__autoskillit__"
-MARKETPLACE_PREFIX = "mcp__plugin_autoskillit_autoskillit__"
-
-
-def _installed_plugins_path() -> Path:
-    """Return the path to Claude Code's installed plugins registry."""
-    return Path.home() / ".claude" / "plugins" / "installed_plugins.json"
-
-
-def detect_autoskillit_mcp_prefix() -> str:
-    """Return the MCP prefix that autoskillit tools will use in a spawned session.
-
-    Reads ~/.claude/plugins/installed_plugins.json to determine whether
-    autoskillit is marketplace-installed. When it is, the marketplace
-    prefix takes precedence even when --plugin-dir is also passed.
-
-    Falls back to DIRECT_PREFIX if the file is absent, unreadable, or
-    does not contain the autoskillit key.
-    """
-    try:
-        data = json.loads(_installed_plugins_path().read_text())
-        if _AUTOSKILLIT_PLUGIN_KEY in data.get("plugins", {}):
-            return MARKETPLACE_PREFIX
-    except (OSError, json.JSONDecodeError, AttributeError, TypeError):
-        pass
-    return DIRECT_PREFIX
+__all__ = [
+    "DIRECT_PREFIX",
+    "MARKETPLACE_PREFIX",
+    "detect_autoskillit_mcp_prefix",
+    "resolve_tool_name",
+]
 
 
 def resolve_tool_name(short_name: str, prefix: str | None = None) -> str:
