+++ src/autoskillit/server/helpers.py
@@ -536,7 +536,8 @@ def _build_hook_diagnostic_warning() -> str | None:
 
 def _get_food_truck_prompt_builder() -> Callable[..., str]:
     """Return the food truck prompt builder with mcp_prefix pre-bound."""
-    from autoskillit.cli import _build_food_truck_prompt, detect_autoskillit_mcp_prefix
+    from autoskillit.core import detect_autoskillit_mcp_prefix
+    from autoskillit.franchise import _build_food_truck_prompt
 
     mcp_prefix = detect_autoskillit_mcp_prefix()
     return functools.partial(_build_food_truck_prompt, mcp_prefix=mcp_prefix)
